LPS-57478 Split out StubMap/StubMapImpl for JDK8 compile time/JDK7 runtime compatibility
